+++ tests/files/issue210.config.flat.js
@@ -1,0 +1,3 @@
+exports.languageOptions = {
+  sourceType: 'module',
+}